Feature/inventory module backend com 7

+++ crm_backend/src/main/scala/order_products/domain/entity/OrderProduct.scala
@@ -3,7 +3,7 @@ package order_products.domain.entity
 case class OrderProduct (
   id: Long = -1,
   productQuantity: Long, 
-  productUnitPrice: Float,
+  productUnitPrice: Double,
   productName: String,
   orderId: Long,
   productLotId: Long
