+++ crm_backend/src/main/scala/sale_products/domain/entity/SaleProduct.scala
@@ -3,10 +3,10 @@ package sale_products.domain.entity
 case class SaleProduct (
   id: Long = -1,
   productQuantity: Double,
-  productDiscount: Float,
+  productDiscount: Double,
   productMeasureUnit: String,
   productUnitPrice: Double,
-  productName: String, 
+  productName: String,
   productDescription: Option[String],
   saleId: Long,
   taxId: Long,
